refactor(theme-utils): update folder architecture
update folder architecture to adhere to our recommended package structure standards

#219
