+++ packages/utils/theme/.npmignore
@@ -1,0 +1,2 @@
+src
+**/*.stories.*
