+++ packages/utils/theme/tsconfig.json
@@ -1,4 +1,4 @@
 {
   "extends": "../../../tsconfig.json",
-  "include": ["**/*", "../../global.d.ts"]
+  "include": ["src/**/*", "../../global.d.ts"]
 }
